chore: rename request-config
