+++ packages/orchestration/src/orchestration-completion-post-request-from-json.test.ts
@@ -1,4 +1,4 @@
-import { constructCompletionPostRequestFromJsonModuleConfig } from './util/orchestration-request-config.js';
+import { constructCompletionPostRequestFromJsonModuleConfig } from './util/orchestration-module-config.js';
 
 describe('construct completion post request from JSON', () => {
   it('should construct completion post request from JSON', () => {
